requirements.txt
Flask
requests
pandas
yfinance
python-dotenv
groq
streamlit